clean up and small fixes

+++ foxspace/time.cpp
@@ -2,16 +2,16 @@
 #include <foxspace/time.hpp>
 #include <thread>
 
-void Fox::sleep_for(unsigned int ms_duration) {
+void Fox::sleep_for(uint32_t ms_duration) {
     std::this_thread::sleep_for(std::chrono::milliseconds(ms_duration));
 }
-unsigned int Fox::get_timepoint() {
+uint64_t Fox::get_timepoint() {
     return std::chrono::time_point_cast<std::chrono::milliseconds>(
                std::chrono::high_resolution_clock::now())
         .time_since_epoch()
         .count();
 }
-Fox::Timer::Timer(unsigned int _duration) : duration{_duration} {
+Fox::Timer::Timer(uint32_t _duration) : duration{_duration} {
     old_time = get_timepoint();
 }
 bool Fox::Timer::done() {
